scripts/requirements.txt
astroid==2.5.6
autopep8==1.5.7
certifi==2020.12.5
chardet==3.0.4
idna==2.8
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
molgenis-py-client==2.3.0
pycodestyle==2.7.0
pylint==2.8.2
python-dotenv==0.17.1
requests==2.21.0
toml==0.10.2
urllib3==1.24.3
wrapt==1.12.1
